// LibrarySystem_FrontEnd/src/app/models/student.model.ts
export class StudentModel {
    accountAddres: string;
    classId: string;
    studentCode: string;
    createTime?: Date;
    email: string;
    id:string;
    identityCardId: string;
    isDelete:number;
    name: string;
    operator: string;
    pictrue: string;
    professionId: string;
    professionName:string;
    remark: string;
    sex: string;
    status: string;
    telNum: string;
    updateTime?: Date;
    institutionName:string;
    classesName:string;
    institutionId:String;

   
}
 