// LibrarySystem_FrontEnd/src/app/models/studentAdd.model.ts
export class StudentAddModel{
    accountAddress: string;
    classesName: string;
    code: string;
    createTime: Date;
    email: string;
    id: string;
    identityCardId: string;
    isDelete: 0;
    name: string;
    operator: string;
    pictrue: string;
    professionName: string;
    remark: string;
    sex: string;
    status: string;
    telNum: string;
    updateTime:Date;
    institutionId:string;
    studentCode:string;
}